requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg[binary]
langchain-core
python-dotenv
sentence-transformers
